# Add perf benchmark and improve perf of <FormattedMessage>