This adds a new `npm run test:perf` script which current does `renderToString()` perf tests.

This also drastically improves the rendering perf of `<FormattedMessage>` in the most common cases.

Here's the results in the `renderToString()` perf tests:

**Before:**
```
100 x <FormattedMessage> x 282 ops/sec ±1.66% (83 runs sampled)
100 x <FormattedMessage> with placeholder x 53.72 ops/sec ±0.73% (69 runs sampled)
```

**After:**
```
100 x <FormattedMessage> x 966 ops/sec ±1.22% (93 runs sampled)
100 x <FormattedMessage> with placeholder x 58.44 ops/sec ±3.16% (61 runs sampled)
```