@ericf 2016-03-07
I updated the benchmark tests to use unique 100 unique messages, which is more accurate than rendering the same message 100 times. This brings down the ops/sec number for the messages with placeholders significantly because it has to create an `IntlMessageFormat` instance per message, and the cache is cleared after every sample run.